+++ tests/test_domain_utils.py
@@ -10,9 +10,9 @@ from ..msticpy.sectools import domain_utils
 
 def test_validate_domain():
     test_dom_val = domain_utils.DomainValidator()
-    valid_tld = test_dom_val.validate_tld("www.goole.com")
-    resolvable = test_dom_val.is_resolvable("www.google.com")
-    blacklisted = test_dom_val.ssl_blacklisted("www.google.com")
+    valid_tld = test_dom_val.validate_tld("www.microsoft.com")
+    resolvable = test_dom_val.is_resolvable("www.microsoft.com")
+    blacklisted = test_dom_val.ssl_blacklisted("www.microsoft.com")
     assert valid_tld == True
     assert resolvable == True
-    assert blacklisted == (False, None)
+    assert blacklisted[0] == False
